refactor: remove unused files, organize src files
