+++ src/components/app/App.tsx
@@ -1,8 +1,12 @@
 import React from 'react';
-import logo from './logo.svg';
+import logo from '../../images/logo.svg';
 import './App.css';
+import { Button } from '@ya.praktikum/react-developer-burger-ui-components';
 
 function App() {
+  fetch('https://norma.nomoreparties.space/api/ingredients')
+    .then((resp) => resp.json())
+    .then((data) => console.log(data));
   return (
     <div className="App">
       <header className="App-header">
@@ -18,6 +22,8 @@ function App() {
         >
           Learn React
         </a>
+        <Button htmlType="button" type="primary" size="large" extraClass="mt-2">Кнопка</Button>
+        <div className="mt-2 text text_type_main-large text_color_inactive">Текст</div>
       </header>
     </div>
   );
